Updating tests and merging
