+++ src/CondenserDotNet.Server/Extensions/ApplicationBuilderExtensions.cs
@@ -7,7 +7,7 @@ namespace CondenserDotNet.Server.Extensions
     {
         public static void UseCondenserRouter(this IApplicationBuilder self)
         {
-            var host = self.ApplicationServices.GetService<RoutingHostLite>();
+            var host = self.ApplicationServices.GetService<RoutingHost>();
             self.UseRouter(host.Router);
         }
     }
